EES-3617 find stats front end tidying
